Fix phase 2 and 3 audit logging in case of internal redirect
Phase 2 and 3 log entries were not logged in the audit log in case of an
internal redirect. Only phase 1 and 4 ones were logged, the former because
early logging was explicitly enabled and the latter because the internal
redirect does not work in this phase.

This commit unconditionally enables early logging in all ModSecurity
phases. Since the Nginx log phase may not be executed in case of an
intervention, the process intervention function is the only place which is
guaranteed to call the log handler in such a case.

Co-authored-by: Dimitris Kolotouros <dkolotouros@noc.grnet.gr>
Co-authored-by: Thanos Giannopoulos <thanosgn@noc.grnet.gr>